Fix `fix` (for real this time?)
Instead of checking for `is_canonical`, only unfold `fix` during `eval`, not
during `quote`. This required switching from a HOAS to a first order
representation of closures, so that we can thread `eval_opts` through.

The `is_canonical` check worked for preventing loops for simple
recursion, but is not sufficient for mutual recursion