elab-letrec-unification: Go back to limiting unfolding with a flag
The previous implementation looped forever when the parameter that
guarded the recursion was not the first parameter in the recursive
function. The strong reduction paper gets around this by having a
multi-parameter fixed-point, where the last parameter guards the
recursion. Because we are implementing an ML/OCaml style language, we
have to use a different approach.