lang-shader-graphics: Fix tests by requiring MDX < 2.4.0
MDX 2.4.0 fails on `lang-shader-graphics/README.md` with:

```
Line 2, characters 28-33:
Error: Unbound module Env
```

It seems like MDX now checks file-part snippets, which is very
frustrating if you want to show a snippet of code with imports and
definitions that lie outside that snippet. I think this new behaviour
might have been introduced in realworldocaml/mdx#446 – the eio
developers seem to have run into this as well.

There doesn’t seem to be a way to turn off this behaviour, so we’ll have
to limit the version range for now.