Cleanups to mutual definition elaboration
This changes functions elaboration (when elaborating recursive
definitions) to take a series of parameter names, as opposed to series
of possibly annotated bindings in order to make the precondition it
expects more understandable. This requires stripping off the annotations
in a separate step, but this should be pretty negligable in terms of
performance impact.